moved fileObject out of __init__ and into its own module
